Rename LaltnLctrl for location.py to AnitiKickWithLocation.py
Kind of a anti kick, it only works when you are NOT on star citzen,  so if you want to afk you can watch something or do what ever, it randomly pulls window to the foreground and returns you to the orginal window you were on, it runs in the back ground on the same timer when in star citizen, between 3-8 minutes. It also logs ur location to a file, buy pressing Lctrl and Lalt it opens chat and sends /showlocation presses enter and closes chat. so you dont have to have chat open all the time. This part is manual, so you have to imitate the code with the keycombo. so its not always "runing" and stoping gameplay to update location. or opening chat.